Implement Portroach::SiteHandler::FTP, HTTP and HttpGuess.
Those handlers are not linked to FindHandler but called direclty from VersionCheck. This is work in progress. Those three handlers take one more argument, $path_ver, which exists if we found a version in the site path.

Each handler return code is now recorded as success or failure.

Checks robotsallowed() and those before verguess() start are still done by VersionCheck as they need more context and this would not fit in the GetFiles() design.